file added

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,8 +10,8 @@
 int main(void)
 {
 	int inc;
-	unsigned long long n1 = 0;
-	unsigned long long n2 = 1, n3;
+	unsigned long n1 = 0;
+	unsigned long n2 = 1, n3;
 	unsigned long n1_h1, n1_h2, n2_h1, n2_h2;
 	for (inc = 0; inc < 98; inc++)
 	{
